Include the Version class in the 64 bit version of base. This is needed so that it can be used from client_utils.cc which is included in both the chrome target and the chrome_nacl_win64 target.
BUG=61609
TEST=None


Review URL: http://codereview.chromium.org/4676002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@65674 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/version.cc
@@ -63,8 +63,8 @@ int Version::CompareTo(const Version& other) const {
 const std::string Version::GetString() const {
   DCHECK(is_valid_);
   std::string version_str;
-  int count = components_.size();
-  for (int i = 0; i < count - 1; ++i) {
+  size_t count = components_.size();
+  for (size_t i = 0; i < count - 1; ++i) {
     version_str.append(base::IntToString(components_[i]));
     version_str.append(".");
   }
